:truck: [#6429] Moving form export logic to import_export module
