+++ src/openforms/forms/admin/form.py
@@ -14,14 +14,14 @@ from ordered_model.admin import OrderedInlineModelAdminMixin, OrderedTabularInli
 
 from openforms.api.utils import underscore_to_camel
 from openforms.emails.models import ConfirmationEmailTemplate
+from openforms.forms.import_export.export_form import export_form
+from openforms.forms.import_export.typing import FormExportOptions
 from openforms.registrations.admin import RegistrationBackendFieldMixin
 from openforms.typing import StrOrPromise
 from openforms.utils.expressions import FirstNotBlank
 
-from ..import_export.typing import FormExportOptions
 from ..models import Category, Form, FormDefinition, FormStep
 from ..models.form import FormsExport
-from ..utils import export_form
 from .mixins import FormioConfigMixin
 from .views import (
     DownloadExportedFormsView,
