+++ src/openforms/forms/constants.py
@@ -3,8 +3,6 @@ from collections.abc import Set
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-EXPORT_META_KEY = "_meta"
-
 
 class LogicActionTypes(models.TextChoices):
     step_not_applicable = (
