+++ src/openforms/forms/import_export/constants.py
@@ -1,0 +1,1 @@
+EXPORT_META_KEY = "_meta"
