+++ src/openforms/forms/tests/factories.py
@@ -4,13 +4,13 @@ import factory.fuzzy
 
 from openforms.authentication.registry import register as authentication_registry
 from openforms.forms.constants import LogicActionTypes
+from openforms.forms.import_export.export_form import form_to_json
 from openforms.products.tests.factories import ProductFactory
 from openforms.registrations.registry import register as registration_registry
 from openforms.variables.constants import FormVariableDataTypes, FormVariableSources
 
 from ..constants import FormTypeChoices, SubmissionAllowedChoices
 from ..models import Form, FormDefinition, FormStep, FormVariable
-from ..utils import form_to_json
 
 
 def authentication_plugins():
